Moved path_events to per_path_events and retrieving path events from the connection
